Title:fix
Description:16157 The menu bar of the deep theater player is always
hidden. You need to click the left button to pause or the right button
to display it

RootCause:Null

Solution:The menu bar of the deep theater player is always hidden. You
need to click the left button to pause or the right button to display it