Title:fix
Description:drag movie item to deepin-movie, ui update would be error
state if drag too fast.

RootCause:Null

Solution: pangguv run on xcb,but the system env var is wayland,add
jugement with xcb or dxcb. if the QT_QPA_PLATFORM config was "xcb",it
would not run with wayland.